requirements.txt
python-chess>=1.10.0
PyQt5>=5.15.10
numpy>=1.26.0
